# apis/bijux-proteomics-runtime/v1/pinned_openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope": {
        "additionalProperties": false,
        "description": "ApiEnvelope.",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunResponse"
              },
              {
                "$ref": "#/components/schemas/RuntimeStatusResponse"
              },
              {
                "$ref": "#/components/schemas/RunArtifactsResponse"
              },
              {
                "$ref": "#/components/schemas/RunEvidenceResponse"
              },
              {
                "$ref": "#/components/schemas/RunReviewResponse"
              },
              {
                "$ref": "#/components/schemas/RunHistoryResponse"
              },
              {
                "$ref": "#/components/schemas/ArtifactLookupResponse"
              },
              {
                "$ref": "#/components/schemas/EvidenceLookupResponse"
              },
              {
                "$ref": "#/components/schemas/InspectResponse"
              },
              {
                "$ref": "#/components/schemas/CompareResponse"
              },
              {
                "$ref": "#/components/schemas/HealthResponse"
              },
              {
                "$ref": "#/components/schemas/ReadyResponse"
              },
              {
                "$ref": "#/components/schemas/RuntimeHealthResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Successful response payload.",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured error payload."
          },
          "meta": {
            "additionalProperties": true,
            "description": "Meta fields.",
            "title": "Meta",
            "type": "object"
          },
          "status": {
            "description": "Response status.",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ApiEnvelope",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "ErrorResponse.",
        "properties": {
          "detail": {
            "description": "Human-readable explanation.",
            "title": "Detail",
            "type": "string"
          },
          "evidence_pointer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional path or identifier that points to relevant evidence.",
            "title": "Evidence Pointer"
          },
          "failure_class": {
            "description": "Stable failure class.",
            "minLength": 1,
            "title": "Failure Class",
            "type": "string"
          },
          "instance": {
            "description": "URI reference for this occurrence.",
            "title": "Instance",
            "type": "string"
          },
          "remediation_hint": {
            "description": "Operator-facing remediation hint.",
            "minLength": 1,
            "title": "Remediation Hint",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "Problem type URI.",
            "format": "uri",
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "instance",
          "failure_class",
          "remediation_hint"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExecutionStatus": {
        "description": "ExecutionStatus.",
        "enum": [
          "completed",
          "errored",
          "aborted"
        ],
        "title": "ExecutionStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunArtifactsResponse": {
        "additionalProperties": false,
        "description": "Stable artifact listing for one run.",
        "properties": {
          "artifacts": {
            "description": "Artifacts attached to the run.",
            "items": {
              "$ref": "#/components/schemas/RuntimeArtifactRecord"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "run_id": {
            "description": "Run identifier.",
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "RunArtifactsResponse",
        "type": "object"
      },
      "RunEvidenceResponse": {
        "additionalProperties": false,
        "description": "Stable evidence-bundle availability response for one run.",
        "properties": {
          "evidence_bundle": {
            "$ref": "#/components/schemas/RuntimeDocumentReference",
            "description": "Evidence-bundle availability."
          },
          "run_id": {
            "description": "Run identifier.",
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "evidence_bundle"
        ],
        "title": "RunEvidenceResponse",
        "type": "object"
      },
...
      "RunResponse": {
        "additionalProperties": false,
        "description": "RunResponse.",
        "properties": {
          "artifacts_dir": {
            "description": "Artifacts directory.",
            "title": "Artifacts Dir",
            "type": "string"
          },
          "candidate_id": {
            "description": "Candidate identifier.",
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "command": {
            "description": "Command name.",
            "title": "Command",
            "type": "string"
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus",
            "description": "Execution status."
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Failure type."
          },
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "provider": {
            "description": "Provider name.",
            "title": "Provider",
            "type": "string"
          },
          "qc_status": {
            "description": "QC status.",
            "title": "Qc Status",
            "type": "string"
          },
          "run_id": {
            "description": "Run identifier.",
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "tool_status": {
            "$ref": "#/components/schemas/ToolStatus",
            "description": "Tool status."
          },
          "version": {
            "$ref": "#/components/schemas/VersionInfo",
            "description": "Version info."
          },
          "warnings": {
            "description": "Warnings.",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "workflow_state": {
            "$ref": "#/components/schemas/WorkflowState",
            "description": "Workflow state."
          }
        },
        "required": [
          "run_id",
          "candidate_id",
          "command",
          "execution_status",
          "workflow_state",
          "outcome",
          "provider",
          "tool_status",
          "qc_status",
          "artifacts_dir",
          "version"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunReviewResponse": {
        "additionalProperties": false,
        "description": "Stable review-packet availability response for one run.",
        "properties": {
          "review_packet": {
            "$ref": "#/components/schemas/RuntimeDocumentReference",
            "description": "Review-packet availability."
          },
          "run_id": {
            "description": "Run identifier.",
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "review_packet"
        ],
        "title": "RunReviewResponse",
        "type": "object"
      },
...
      "RuntimeStatusResponse": {
        "additionalProperties": false,
        "description": "Stable runtime status response for one run.",
        "properties": {
          "evidence_bundle": {
            "$ref": "#/components/schemas/RuntimeDocumentReference",
            "description": "Evidence-bundle availability for the run."
          },
          "review_packet": {
            "$ref": "#/components/schemas/RuntimeDocumentReference",
            "description": "Review-packet availability for the run."
          },
          "summary": {
            "$ref": "#/components/schemas/RunResponse",
            "description": "Canonical run summary."
          }
        },
        "required": [
          "summary",
          "evidence_bundle",
          "review_packet"
        ],
        "title": "RuntimeStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API exposing the same capabilities as the CLI, nothing more.",
    "title": "bijux-proteomics-runtime",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifacts": {
      "get": {
        "description": "Return the stable artifact lookup response.",
        "operationId": "artifact_lookup_endpoint_api_v1_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "artifact_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Kind"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_query_cost",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Max Query Cost",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Artifact Lookup Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
...
    "/api/v1/evidence": {
      "get": {
        "description": "Return the stable evidence and review lookup response.",
        "operationId": "evidence_lookup_endpoint_api_v1_evidence_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "document_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Kind"
            }
          },
          {
            "in": "query",
            "name": "availability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Availability"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_query_cost",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Max Query Cost",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Evidence Lookup Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
...
    "/api/v1/runs/history": {
      "get": {
        "description": "Return the stable run-history lookup response.",
        "operationId": "run_history_endpoint_api_v1_runs_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "workflow_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow State"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_query_cost",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Max Query Cost",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Run History Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
    "/api/v1/runs/{run_id}/artifacts": {
      "get": {
        "description": "Return the stable artifact inventory for one run.",
        "operationId": "run_artifacts_endpoint_api_v1_runs__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Artifacts Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
    "/api/v1/runs/{run_id}/evidence-bundle": {
      "get": {
        "description": "Return the stable evidence-bundle surface for one run.",
        "operationId": "run_evidence_endpoint_api_v1_runs__run_id__evidence_bundle_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_document",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Document",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_inline_bytes",
            "required": false,
            "schema": {
              "default": 256000,
              "title": "Max Inline Bytes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evidence Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
    "/api/v1/runs/{run_id}/review-packet": {
      "get": {
        "description": "Return the stable review-packet surface for one run.",
        "operationId": "run_review_endpoint_api_v1_runs__run_id__review_packet_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_document",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Document",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_inline_bytes",
            "required": false,
            "schema": {
              "default": 256000,
              "title": "Max Inline Bytes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Review Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
    "/api/v1/runs/{run_id}/status": {
      "get": {
        "description": "Return the stable runtime status contract for one run.",
        "operationId": "run_status_endpoint_api_v1_runs__run_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_documents",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Documents",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_inline_bytes",
            "required": false,
            "schema": {
              "default": 256000,
              "title": "Max Inline Bytes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Run Status Endpoint",
        "tags": [
          "runtime-contracts"
        ]
      }
    },
...
  }
}